requirements.txt
Tornado>=3
Mongoengine
unidecode
py-bcrypt
numpy
